removed main window from starting method to make future refactoring easier
